requirements.txt
gradio
requests
langchain
langchain-community
langchain-core
langchain-openai
langchain-google-genai
langchain-tavily
langgraph
wikipedia
python-dotenv
pydantic
annotated-types
anyio
uvicorn
requests
fastapi
pydantic-settings